Refactor building simulation maintenance and production modules
