+++ packages/engine/src/simulation/eventPipeline.ts
@@ -1,6 +1,6 @@
 import type { SimResources } from '../index';
 import type { GameTime } from '../types/gameTime';
-import type { SimulatedBuilding } from './buildingSimulation';
+import type { SimulatedBuilding } from './buildings';
 import type { Citizen } from './citizens/citizen';
 import type { WorkerProfile } from './workers/types';
 import { EventManager } from './events/EventManager';
